Implement concept of partitions.
A partition is a set of transactions in the mempool that are related to each other.  All tx's in the mempool belong to exactly one partition and all tx's that are reachable from that tx by looking at other mempool tx's that are TxIn.prevout's (parents) or spend TxOut's (children) are in the same partition.  This is the transitive closure of transactions in the mempool.