Fix minor bugs in miner_tests
There was an extraneous tx being added to the mempool which spent the same inputs.  This only didn't cause a problem because locktimes stopped it from being selected for the template.

In addition, since the merger of MedianTimePast, locked txs that were time based were not being correctly tested in the unit test.