Use a persistent CCoinsViewCache, tipCache, for connecting blocks.
Keep tipCache warm by populating it every 30 seconds with inputs required for the block that would be created from the mempool.

WARNING: More work needs to be put into place to handle consistency of tipCache.  In particular there are other failure modes of ConnectBlock and DisconnectBlock such as in VerifyDB or TestBlockValidity that aren't address.  And the existing cases are kind of hacked in.  Perhaps RAII would work?