Prefer any answer to no answer
If any of the time horizons in estimateCombinedFee can return an answer, always use this rather than returning a possible -1.

Additionally, if we are not using conservative estimates, but we get a -1 use the conservative method of checking longer time horizons to try to obtain an answer for 2*target.  This might happen if the node has been down for a little while and short term estimates have decayed.